some cleanup/reorg
